requirements.txt
bs4==0.0.1
Django==2.2.3
djangorestframework==3.10.2